+++ CHANGELOG.md
@@ -180,6 +180,7 @@ batch
 - [#584](https://github.com/influxdata/kapacitor/issues/584): Do not block during startup to send usage stats.
 - [#553](https://github.com/influxdata/kapacitor/issues/553): Periodically check if new InfluxDB DBRPs have been created.
 - [#602](https://github.com/influxdata/kapacitor/issues/602): Fix missing To property on email alert handler.
+- [#581](https://github.com/influxdata/kapacitor/issues/581): Record/Replay batch tasks get cluster info from task not API.
 
 ## v0.13.1 [2016-05-13]